+++ src/main/java/net/synchlabs/backrooms/client/BackroomsClient.java
@@ -8,6 +8,7 @@ import net.minecraft.client.render.SkyProperties;
 import net.minecraft.util.Identifier;
 import net.synchlabs.backrooms.init.BackroomsBlocks;
 import net.synchlabs.backrooms.world.Level0;
+import net.synchlabs.backrooms.world.Level0Sky;
 
 import java.util.Map;
 
@@ -21,7 +22,7 @@ public class BackroomsClient implements ClientModInitializer, ExtraSkiesApi {
 
 	@Override
 	public void registerModSkies(Map<Identifier, SkyProperties> map) {
-		SkyProperties Level0Sky = new Level0.Level0Sky();
+		SkyProperties Level0Sky = new Level0Sky();
 		map.put(Level0.LEVEL_0_ID,Level0Sky);
 	}
 }
