+++ src/main/java/net/synchlabs/backrooms/world/Level0.java
@@ -45,27 +45,6 @@ public class Level0 extends ExtraDimension{
 
 	}
 
-	public static class Level0Sky extends SkyProperties {
 
-		public Level0Sky() {
-			super(Float.NaN, false, SkyType.NONE, false, false);
-		}
-
-		@Override
-		public Vec3d adjustFogColor(Vec3d color, float sunHeight) {
-			return color;
-		}
-
-		@Override
-		public boolean useThickFog(int camX, int camY) {
-			return false;
-		}
-
-		@Override
-		public float[] getFogColorOverride(float skyAngle, float tickDelta) {
-			return null;
-		}
-
-	}
 
 }
